* winsup.api/crlf.c: Fix C signed/unsigned compare warning. 	* winsup.api/mmaptest01.c: Ditto. 	* winsup.api/ltp/chmod01.c: Ditto. 	* winsup.api/ltp/fork04.c: Ditto. 	* winsup.api/ltp/lseek03.c: Ditto. 	* winsup.api/ltp/lseek06.c: Ditto. 	* winsup.api/ltp/lseek07.c: Ditto. 	* winsup.api/ltp/lseek08.c: Ditto. 	* winsup.api/ltp/mmap001.c: Ditto. 	* winsup.api/ltp/mmap02.c: Ditto. 	* winsup.api/ltp/mmap03.c: Ditto. 	* winsup.api/ltp/mmap04.c: Ditto. 	* winsup.api/ltp/mmap05.c: Ditto. 	* winsup.api/ltp/mmap06.c: Ditto. 	* winsup.api/ltp/mmap07.c: Ditto. 	* winsup.api/ltp/mmap08.c: Ditto. 	* winsup.api/ltp/pipe11.c: Ditto. 	* winsup.api/ltp/poll01.c: Ditto. 	* winsup.api/ltp/sync02.c: Ditto. 	* winsup.api/ltp/times03.c: Ditto. 	* winsup.api/ltp/umask03.c: Ditto. 	* winsup.api/ltp/getpgid01.c: Remove unused obsolete include. 	* winsup.api/ltp/getpgid02.c: Ditto.

+++ winsup/testsuite/winsup.api/ltp/fork04.c
@@ -30,7 +30,7 @@
  * http://oss.sgi.com/projects/GenInfo/NoticeExplan/
  *
  */
-/* $Id: fork04.c,v 1.3 2003/01/24 01:09:39 cgf Exp $ */
+/* $Id: fork04.c,v 1.4 2003/02/20 11:22:17 corinna Exp $ */
 /**********************************************************
  * 
  *    OS Test - Silicon Graphics, Inc.
@@ -129,7 +129,7 @@ extern int Tst_count;		/* Test Case counter for tst_* routines */
 
 /* list of environment variables to test */
 const char *environ_list[] = {"TERM","NoTSetzWq","TESTPROG"};
-#define NUMBER_OF_ENVIRON sizeof(environ_list)/sizeof(char *)
+#define NUMBER_OF_ENVIRON (int)(sizeof(environ_list)/sizeof(char *))
 int TST_TOTAL=NUMBER_OF_ENVIRON;		/* Total number of test cases. */
 
 /***************************************************************
